The `submit-nuget` (Automatic Dependency Submission) job failed because
GitHub's dynamic NuGet workflow tried to checkout branch
`alert-autofix-2262` — a GitHub Advanced Security autofix branch for
CodeQL alert #2262 — which had already been deleted.

## Changes

- **`pxt_modules/mixer/melody.cpp`** — Fix CodeQL
`cpp/multiplication-result-converted-to-larger-type` (alert #2262):
`numsamples * 2` was computed as `int` then widened to `size_t` for
`memset`, risking silent overflow. Cast before multiplying:
  ```cpp
  // Before
  memset(dst, 0, numsamples * 2);
  // After
  memset(dst, 0, (size_t)numsamples * 2);
  ```

- **`pxt_modules/settings/RAFFS.cpp`** — Restore `static uint16_t
raffs_unlocked_event` declaration and its constructor initialization
(`codal::allocateNotifyEvent()`), which were removed by a prior autofix.
The variable is referenced in `FS::lock()` and `FS::unlock()` and must
be defined at file scope.

+++ pxt_modules/mixer/melody.cpp
@@ -243,7 +243,7 @@ int WSynthesizer::fillSamples(int16_t *dst, int numsamples) {
     return 1;
   }
 
-  memset(dst, 0, numsamples * 2);
+  memset(dst, 0, (size_t)numsamples * 2);
 
   uint32_t samplesPerMS = (sampleRate << 8) / 1000;
   float toneStepMult = (1024.0 * (1 << 16)) / sampleRate;
